#1010 [REFACTOR] feature에 report 기능 폴더로 분리
